# speedup scanning by 3.7x using gitoxide