--- a/src/blob_id.rs
+++ b/src/blob_id.rs
@@ -1,4 +1,5 @@
 use anyhow::Result;
+use git_repository as git;
 
 // -------------------------------------------------------------------------------------------------
 // BlobId
@@ -57,6 +58,16 @@ impl BlobId {
     #[inline]
     pub fn bytes(&self) -> &[u8] {
         &self.0
+    }
+}
+
+impl<'a> From<&'a git::ObjectId> for BlobId {
+    fn from(id: &'a git::ObjectId) -> Self {
+        BlobId(
+            id.as_bytes()
+                .try_into()
+                .expect("oid should be a 20-byte value"),
+        )
     }
 }