This implementation is the minimal changes required to use `gitoxide` instead of `git2` to extract
blob data.

Please note that the first commit is to be able to build on ARM, effectively disabling the actual scanning.
With some work, I think the rules database could be made to run with [`RegexSet`](https://github.com/rust-lang/regex#usage-match-multiple-regular-expressions-simultaneously)
making this useful tool more accessible.

Lastly, if the scanning step would be skipped, one could go straight to decoding packs directly to not waste a cycle on decoding packed objects, further speeding up decoding
to reach light speed (i.e., the maximum possible speed :D).

### Results

* 1.46 GiB/s) GiB/s `git2` blob extraction
* ~~4.67 GiB/s `gitoxide` blob extraction (3.2x)~~
* 5.47GiB/s `gitoxide` blob extraction (3.7x)

The speedup above will be even more significant on machines with more cores - `git2` has globally locked state somewhere and can't scale well even when opening multiple repository handles.

Note that the above numbers are created on ARM and I had to disable the actual scanning. There are still ways to speed this up. Trivially by avoiding to clone the blob data, for a few percent maybe, and more radically by changing the algorithm to leverage `gitoxide` pack resolution, which can bring up data decompression performance to 12GB/s on my machine for the kernel pack, and I have seen 36GB/s on a Ryzen.

That means in theory, if scanning is free, we are looking at 2.5s for scanning the entire linux kernel (on a Ryzen).

<details>

```
❯ cargo run --release  --no-default-features  -- scan  -d datastore-delme ~/dev/github.com/git/git/
    Finished release [optimized] target(s) in 0.08s
     Running `target/release/noseyparker scan -d datastore-delme /Users/byron/dev/github.com/git/git/`
Found 17.06 GiB from 6,056 plain files and 535,038 blobs from 5 Git repos [00:00:02]
Scanning content  ████████████████████ 100%  17.06 GiB/17.06 GiB  [00:00:04]                                                                                                                                                 Scanned 6.16 GiB from 168,667 blobs in 4 seconds (1.46 GiB/s); 0/0 new matches

Run the `report` command next to show finding details.

noseyparker ( gitoxide) took 7s
❯ cargo run --release  --no-default-features  -- scan  -d datastore-delme ~/dev/github.com/git/git/
   Compiling once_cell v1.16.0
   Compiling smallvec v1.10.0
   Compiling lock_api v0.4.9
   Compiling parking_lot_core v0.9.5
   Compiling thiserror-impl v1.0.37
   Compiling cmake v0.1.49
   Compiling quick-error v2.0.1
   Compiling num-traits v0.2.15
   Compiling crossbeam-queue v0.3.8
   Compiling sha1-asm v0.5.1
   Compiling crc32fast v1.3.2
   Compiling bstr v1.0.1
   Compiling adler v1.0.2
   Compiling libz-sys v1.1.8
   Compiling miniz_oxide v0.6.2
   Compiling ahash v0.7.6
   Compiling human_format v1.0.3
   Compiling thiserror v1.0.37
   Compiling bytesize v1.1.0
   Compiling time-core v0.1.0
   Compiling hashbrown v0.12.3
   Compiling num_threads v0.1.6
   Compiling time-macros v0.2.6
   Compiling crossbeam v0.8.2
   Compiling prodash v21.1.0
   Compiling jwalk v0.6.0
   Compiling sha1_smol v1.0.0
   Compiling parking_lot v0.12.1
   Compiling minimal-lexical v0.2.1
   Compiling signal-hook v0.3.14
   Compiling signal-hook-registry v1.4.0
   Compiling serde v1.0.147
   Compiling git-hash v0.10.1
   Compiling git-validate v0.7.0
   Compiling nom v7.1.1
   Compiling git-path v0.6.0
   Compiling fastrand v1.8.0
   Compiling btoi v0.4.2
   Compiling time v0.3.17
   Compiling remove_dir_all v0.5.3
   Compiling tempfile v3.3.0
   Compiling dashmap v5.4.0
   Compiling memmap2 v0.5.8
   Compiling hash_hasher v2.0.3
   Compiling ahash v0.8.2
   Compiling rustversion v1.0.9
   Compiling git-quote v0.4.0
   Compiling git-tempfile v3.0.0
   Compiling git-config-value v0.9.0
   Compiling git-glob v0.5.0
   Compiling git-sec v0.5.0
   Compiling git-lock v3.0.0
   Compiling imara-diff v0.1.5
   Compiling unicode-bom v1.1.4
   Compiling git-date v0.3.0
   Compiling arrayvec v0.7.2
   Compiling git-actor v0.14.1
   Compiling castaway v0.2.2
   Compiling atoi v1.0.0
   Compiling git-chunk v0.4.0
   Compiling uluru v3.0.0
   Compiling git-command v0.2.0
   Compiling nix v0.25.1
   Compiling tracing-core v0.1.30
   Compiling git-bitmap v0.2.0
   Compiling compact_str v0.6.1
   Compiling filetime v0.2.19
   Compiling clru v0.5.0
   Compiling home v0.5.4
   Compiling thread_local v1.1.4
   Compiling io-close v0.3.7
   Compiling utf8-width v0.1.6
   Compiling arc-swap v1.5.1
   Compiling git-prompt v0.2.0
   Compiling tracing-log v0.1.3
   Compiling indexmap v1.9.2
   Compiling git-mailmap v0.6.0
   Compiling hashlink v0.8.1
   Compiling tracing v0.1.37
   Compiling clap v4.0.29
   Compiling openssl v0.10.43
   Compiling tracing-subscriber v0.3.16
   Compiling rusqlite v0.28.0
   Compiling sha1 v0.10.5
   Compiling bstr v0.2.17
   Compiling byte-unit v4.0.17
   Compiling serde_json v1.0.88
   Compiling serde_yaml v0.9.14
   Compiling globset v0.4.9
   Compiling csv v1.1.6
   Compiling prettytable-rs v0.9.0
   Compiling ignore v0.4.18
   Compiling libssh2-sys v0.2.23
   Compiling flate2 v1.0.25
   Compiling git-features v0.24.1
   Compiling git-object v0.23.0
   Compiling git-attributes v0.6.0
   Compiling git-url v0.11.0
   Compiling libgit2-sys v0.14.0+1.5.0
   Compiling git-credentials v0.7.0
   Compiling git-traverse v0.19.0
   Compiling git-ref v0.20.0
   Compiling git-diff v0.23.0
   Compiling git-revision v0.7.0
   Compiling git-pack v0.27.0
   Compiling git-index v0.9.1
   Compiling git-refspec v0.4.0
   Compiling git-config v0.12.0
   Compiling git-worktree v0.9.0
   Compiling git-discover v0.9.0
   Compiling git-odb v0.37.0
   Compiling git-repository v0.29.0
   Compiling git2 v0.15.0
   Compiling noseyparker v0.10.0 (/Users/byron/dev/github.com/praetorian-inc/noseyparker)
    Finished release [optimized] target(s) in 25.99s
     Running `target/release/noseyparker scan -d datastore-delme /Users/byron/dev/github.com/git/git/`
Found 17.06 GiB from 6,056 plain files and 535,038 blobs from 5 Git repos [00:00:02]
Scanning content  ████████████████████ 100%  17.06 GiB/17.06 GiB  [00:00:01]                                                                                                                                                 Scanned 6.12 GiB from 165,099 blobs in 1 second (4.67 GiB/s); 0/0 new matches

Run the `report` command next to show finding details.

noseyparker ( gitoxide) +1081 -62 [!] took 29s
```
</details>