@bradlarsen 2022-12-09
Hi @Byron. This is terrific; thank you for submitting this PR!

I had seen `gitoxide` previously, and considered it for reading Git objects instead of `git2`, but had not had a chance to investigate further. The use of `git2` is currently a performance bottleneck when scanning, particularly when using many parallel workers.

Let me kick the tires on this a bit. I have a corpus of over 7000 Git repos that I'd like to run these changes on.