@bradlarsen 2022-12-09
> With some work, I think the rules database could be made to run with [RegexSet](https://github.com/rust-lang/regex#usage-match-multiple-regular-expressions-simultaneously) making this useful tool more accessible.

Indeed, the current version of Nosey Parker requires an x86_64 CPU to easily build and run.

It is not impossible to run Hyperscan on ARM, however. There is a fork that adds support for additional platforms: https://github.com/vectorcamp/vectorscan. I have some local changes to use that fork instead of Hyperscan, and it seems to work on a MacBook Pro with an M1 Max. I'm hoping to get these changes merged into `main` in the next week or so.

I have not tried `RegexSet` specifically. But in an earlier implementation of Nosey Parker, I did briefly try [`re2`](https://github.com/google/re2), which similarly supports scanning for many patterns simultaneously. That experiment resulted in an order of magnitude slower scanning. I suspect `RegexSet` would also be slower than Hyperscan (which at least with certain rulesets can scan at >25Gbps on a single core).

The experiment using `RegexSet` would be interesting, but at this point we will probably be sticking with Hyperscan (and soon moving to the vectorscan fork that supports ARM).