@bradlarsen 2022-12-09
> Vectorscan sounds like the way to go on M1 and I am really looking forward to running it for the first time on my own machine proper, with an actual scan. Then it should become clear if it’s worth to get this sped up to 12GB /s.

If you are impatient and ambitious, you could build with vectorscan yourself, as I sketched in #5. Big idea: build vectorscan from source, set the `HYPERSCAN_ROOT` environment variable appropriately, then `cargo build`, and it should work.