@bradlarsen 2022-12-12
@Byron I have run some tests with the modified blob-reading code from this PR that uses `gitoxide` instead of `git2`. 
I want this in Nosey Parker.

I ran with actual scanning via Hyperscan still enabled. I see a 2-2.5x overall speedup from your changes from a few repos I tried. For example, when running over 100GB of Linux history on my 10-core M1 Max: 2.51GiB/s scan rate with your changes vs 1.18GiB/s without.

I am still in process of validating this with my corpus of 7000+ Git repos. I suspect it will run 2-3x faster in total with your changes. The speed is great; I just want to battle-test it some more.

A question for you from your original description:

> if the scanning step would be skipped, one could go straight to decoding packs directly to not waste a cycle on decoding packed objects, further speeding up decoding to reach light speed

What do you mean here—can you expand on this? Are there particular gitoxide crates or APIs I should take a look at?