@bradlarsen 2022-12-12
> changing the algorithm to leverage gitoxide pack resolution

@Byron can you also please elaborate a little more what you have in mind here? Are there crates or APIs I should be looking at?