@Byron 2022-12-13
That's fantastic news! I'd be keen to hear the final numbers to learn how long it takes to go through all these 7k repositories, and maybe how much data was churned. Again, depending on the pack compression and with the right machine 35GB/s are achievable.

To do that, you'd have to use a special cache that accelerates decoding of packs. There is `gix verify` which I believe already [decodes each and every object](https://github.com/Byron/gitoxide/blob/b77a3903300442dc10ca447e62016aed2ce3f2a7/gitoxide-core/src/repository/verify.rs#L35-L45) of a repository with data rates as high as 12GB/s on the linux kernel on M1. Doing it like this requires some changes to the core algorithm, but the linked code should show how it's done.

Lastly, the enumeration step is something I'd make optional, as instead you can rely and visualize the progress information that `gitoxide` delivers. There is [a portion in the upcoming `cargo`](https://github.com/Byron/cargo/blob/e8c4201531038b39d447c91a56a6e4a5a2708bf2/src/cargo/sources/git/oxide.rs#L45-L109) that shows how to feed your own progress bars with it. With prior enumeration you will have bounds for progress bars, but these are expensive numbers.

I have also opened [a tracking issue](https://github.com/Byron/gitoxide/issues/654) with features I think you will need to fully replace `git2` in this codebase.

I hope that helps.